Add files via upload
This is a prototype. The goal of the game is to generate circles of random size and color in random locations on the screen. The circles should disappear when touched. However I realized generating and removing a new SKShapeNode each time was inefficient so I'll be changing my approach by generating a set of SKShapeNode's in the beginning and alternating between them.